+++ bootstrap/app.php
@@ -84,6 +84,7 @@ $container['view'] = function ($container) use (&$config) {
 	$view->getEnvironment()->addFunction(new Twig_Function('lang', 'lang'));
 	$view->getEnvironment()->addFunction(new Twig_Function('urlFor', 'urlFor'));
 	$view->getEnvironment()->addFunction(new Twig_Function('mime2font', 'mime2font'));
+	$view->getEnvironment()->addFunction(new Twig_Function('queryParams', 'queryParams'));
 	return $view;
 };